+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/service/TriggerInfoMaintainServiceImpl.java
@@ -2,7 +2,7 @@ package com.dwarfeng.fdr.impl.service;
 
 import com.dwarfeng.fdr.stack.bean.entity.TriggerInfo;
 import com.dwarfeng.fdr.stack.service.TriggerInfoMaintainService;
-import com.dwarfeng.subgrade.impl.service.CustomCrudService;
+import com.dwarfeng.subgrade.impl.service.CustomBatchCrudService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
@@ -14,11 +14,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Service
 public class TriggerInfoMaintainServiceImpl implements TriggerInfoMaintainService {
 
     @Autowired
-    private CustomCrudService<LongIdKey, TriggerInfo> crudDelegate;
+    private CustomBatchCrudService<LongIdKey, TriggerInfo> crudService;
     @Autowired
     private DaoOnlyEntireLookupService<TriggerInfo> entireLookupService;
     @Autowired
@@ -28,70 +30,148 @@ public class TriggerInfoMaintainServiceImpl implements TriggerInfoMaintainServic
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
     public boolean exists(LongIdKey key) throws ServiceException {
-        return crudDelegate.exists(key);
+        return crudService.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
     public TriggerInfo get(LongIdKey key) throws ServiceException {
-        return crudDelegate.get(key);
+        return crudService.get(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
     public LongIdKey insert(TriggerInfo element) throws ServiceException {
-        return crudDelegate.insert(element);
+        return crudService.insert(element);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
     public void update(TriggerInfo element) throws ServiceException {
-        crudDelegate.update(element);
+        crudService.update(element);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
     public void delete(LongIdKey key) throws ServiceException {
-        crudDelegate.delete(key);
+        crudService.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
     public TriggerInfo getIfExists(LongIdKey key) throws ServiceException {
-        return crudDelegate.getIfExists(key);
+        return crudService.getIfExists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
     public LongIdKey insertIfNotExists(TriggerInfo element) throws ServiceException {
-        return crudDelegate.insertIfNotExists(element);
+        return crudService.insertIfNotExists(element);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
     public void updateIfExists(TriggerInfo element) throws ServiceException {
-        crudDelegate.updateIfExists(element);
+        crudService.updateIfExists(element);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
     public void deleteIfExists(LongIdKey key) throws ServiceException {
-        crudDelegate.deleteIfExists(key);
+        crudService.deleteIfExists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
     public LongIdKey insertOrUpdate(TriggerInfo element) throws ServiceException {
-        return crudDelegate.insertOrUpdate(element);
+        return crudService.insertOrUpdate(element);
+    }
+
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public boolean allExists(List<LongIdKey> keys) throws ServiceException {
+        return crudService.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public boolean nonExists(List<LongIdKey> keys) throws ServiceException {
+        return crudService.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public List<TriggerInfo> batchGet(List<LongIdKey> keys) throws ServiceException {
+        return crudService.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public List<LongIdKey> batchInsert(List<TriggerInfo> elements) throws ServiceException {
+        return crudService.batchInsert(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void batchUpdate(List<TriggerInfo> elements) throws ServiceException {
+        crudService.batchUpdate(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void batchDelete(List<LongIdKey> keys) throws ServiceException {
+        crudService.batchDelete(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public List<TriggerInfo> batchGetIfExists(List<LongIdKey> keys) throws ServiceException {
+        return crudService.batchGetIfExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public List<LongIdKey> batchInsertIfExists(List<TriggerInfo> elements) throws ServiceException {
+        return crudService.batchInsertIfExists(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void batchUpdateIfExists(List<TriggerInfo> elements) throws ServiceException {
+        crudService.batchUpdateIfExists(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void batchDeleteIfExists(List<LongIdKey> keys) throws ServiceException {
+        crudService.batchDeleteIfExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public List<LongIdKey> batchInsertOrUpdate(List<TriggerInfo> elements) throws ServiceException {
+        return crudService.batchInsertOrUpdate(elements);
     }
 
     @Override