+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/service/operation/FilterInfoCrudOperation.java
@@ -5,17 +5,19 @@ import com.dwarfeng.fdr.stack.cache.EnabledFilterInfoCache;
 import com.dwarfeng.fdr.stack.cache.FilterInfoCache;
 import com.dwarfeng.fdr.stack.dao.FilterInfoDao;
 import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes;
-import com.dwarfeng.subgrade.sdk.service.custom.operation.CrudOperation;
+import com.dwarfeng.subgrade.sdk.service.custom.operation.BatchCrudOperation;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Objects;
 
 @Component
-public class FilterInfoCrudOperation implements CrudOperation<LongIdKey, FilterInfo> {
+public class FilterInfoCrudOperation implements BatchCrudOperation<LongIdKey, FilterInfo> {
 
     @Autowired
     private FilterInfoDao filterInfoDao;
@@ -81,5 +83,52 @@ public class FilterInfoCrudOperation implements CrudOperation<LongIdKey, FilterI
 
         filterInfoDao.delete(key);
         filterInfoCache.delete(key);
+    }
+
+    @Override
+    public boolean allExists(List<LongIdKey> keys) throws Exception {
+        return filterInfoCache.allExists(keys) || filterInfoDao.allExists(keys);
+    }
+
+    @Override
+    public boolean nonExists(List<LongIdKey> keys) throws Exception {
+        return filterInfoCache.nonExists(keys) && filterInfoCache.nonExists(keys);
+    }
+
+    @Override
+    public List<FilterInfo> batchGet(List<LongIdKey> keys) throws Exception {
+        if (filterInfoCache.allExists(keys)) {
+            return filterInfoCache.batchGet(keys);
+        } else {
+            if (!filterInfoDao.allExists(keys)) {
+                throw new ServiceException(ServiceExceptionCodes.ENTITY_NOT_EXIST);
+            }
+            List<FilterInfo> filterInfos = filterInfoDao.batchGet(keys);
+            filterInfoCache.batchPush(filterInfos, filterInfoTimeout);
+            return filterInfos;
+        }
+    }
+
+    @Override
+    public List<LongIdKey> batchInsert(List<FilterInfo> filterInfos) throws Exception {
+        List<LongIdKey> keys = new ArrayList<>();
+        for (FilterInfo filterInfo : filterInfos) {
+            keys.add(insert(filterInfo));
+        }
+        return keys;
+    }
+
+    @Override
+    public void batchUpdate(List<FilterInfo> filterInfos) throws Exception {
+        for (FilterInfo filterInfo : filterInfos) {
+            update(filterInfo);
+        }
+    }
+
+    @Override
+    public void batchDelete(List<LongIdKey> keys) throws Exception {
+        for (LongIdKey key : keys) {
+            delete(key);
+        }
     }
 }