+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/service/TriggerInfoMaintainService.java
@@ -2,7 +2,7 @@ package com.dwarfeng.fdr.stack.service;
 
 import com.dwarfeng.fdr.stack.bean.entity.TriggerInfo;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
-import com.dwarfeng.subgrade.stack.service.CrudService;
+import com.dwarfeng.subgrade.stack.service.BatchCrudService;
 import com.dwarfeng.subgrade.stack.service.EntireLookupService;
 import com.dwarfeng.subgrade.stack.service.PresetLookupService;
 
@@ -12,7 +12,7 @@ import com.dwarfeng.subgrade.stack.service.PresetLookupService;
  * @author DwArFeng
  * @since 0.0.1-alpha
  */
-public interface TriggerInfoMaintainService extends CrudService<LongIdKey, TriggerInfo>,
+public interface TriggerInfoMaintainService extends BatchCrudService<LongIdKey, TriggerInfo>,
         EntireLookupService<TriggerInfo>, PresetLookupService<TriggerInfo> {
 
     String CHILD_FOR_POINT = "child_for_point";